refactored code, tested ngspice and xyce on examples
